+++ src/Cli/dotnet/Commands/Test/MTP/Terminal/TerminalTestReporter.cs
@@ -854,41 +854,6 @@ internal sealed partial class TerminalTestReporter : IDisposable
         });
     }
 
-    public void WriteMessage(string text, SystemConsoleColor? color = null, int? padding = null)
-    {
-        if (color != null)
-        {
-            _terminalWithProgress.WriteToTerminal(terminal =>
-            {
-                terminal.SetColor(ToTerminalColor(color.ConsoleColor));
-                if (padding == null)
-                {
-                    terminal.AppendLine(text);
-                }
-                else
-                {
-                    AppendIndentedLine(terminal, text, new string(' ', padding.Value));
-                }
-
-                terminal.ResetColor();
-            });
-        }
-        else
-        {
-            _terminalWithProgress.WriteToTerminal(terminal =>
-            {
-                if (padding == null)
-                {
-                    terminal.AppendLine(text);
-                }
-                else
-                {
-                    AppendIndentedLine(terminal, text, new string(' ', padding.Value));
-                }
-            });
-        }
-    }
-
     internal void TestDiscovered(
         string executionId,
         string? displayName,
@@ -998,123 +963,5 @@ internal sealed partial class TerminalTestReporter : IDisposable
         }
 
         _terminalWithProgress.UpdateWorker(asm.SlotIndex);
-    }
-
-    public void WritePlatformAndExtensionOptions(HelpContext context,
-        IEnumerable<CommandLineOptionMessage> builtInOptions,
-        IEnumerable<CommandLineOptionMessage> nonBuiltInOptions,
-        Dictionary<bool, List<(string[], string[])>> moduleToMissingOptions)
-    {
-        if (_wasCancelled)
-        {
-            return;
-        }
-
-        if (builtInOptions.Any())
-        {
-            WriteOtherOptionsSection(context, CliCommandStrings.HelpPlatformOptions, builtInOptions);
-            context.Output.WriteLine();
-        }
-
-        if (nonBuiltInOptions.Any())
-        {
-            WriteOtherOptionsSection(context, CliCommandStrings.HelpExtensionOptions, nonBuiltInOptions);
-            context.Output.WriteLine();
-        }
-        WriteModulesToMissingOptionsToConsole(moduleToMissingOptions);
-    }
-
-    private void WriteOtherOptionsSection(HelpContext context, string title, IEnumerable<CommandLineOptionMessage> options)
-    {
-        List<TwoColumnHelpRow> optionRows = [];
-
-        foreach (var option in options)
-        {
-            if (option.IsHidden != true)
-            {
-                optionRows.Add(new TwoColumnHelpRow($"--{option.Name}", option.Description ?? string.Empty));
-            }
-        }
-
-        if (optionRows.Count > 0)
-        {
-            WriteHeading(title, null);
-            context.HelpBuilder.WriteColumns(optionRows, context);
-        }
-    }
-
-
-    private void WriteHeading(string? heading, string? description)
-    {
-        if (!string.IsNullOrWhiteSpace(heading))
-        {
-            WriteMessage(heading);
-        }
-
-        if (!string.IsNullOrWhiteSpace(description))
-        {
-            int maxWidth = int.MaxValue - SingleIndentation.Length;
-            foreach (var part in WrapText(description!, maxWidth))
-            {
-                WriteMessage(SingleIndentation);
-                WriteMessage(part);
-            }
-        }
-    }
-
-    private static IEnumerable<string> WrapText(string text, int maxWidth)
-    {
-        if (string.IsNullOrWhiteSpace(text))
-        {
-            yield break;
-        }
-
-        foreach (var part in text.Split(["\r\n", "\n"], StringSplitOptions.None))
-        {
-            for (int i = 0; i < part.Length; i += maxWidth)
-            {
-                int length = Math.Min(maxWidth, part.Length - i);
-                int lastSpace = part.LastIndexOf(' ', i + length, length);
-                if (lastSpace > i)
-                {
-                    length = lastSpace - i + 1;
-                }
-                yield return part.Substring(i, length).TrimEnd();
-            }
-        }
-    }
-
-    public void WriteModulesToMissingOptionsToConsole(Dictionary<bool, List<(string[], string[])>> modulesWithMissingOptions)
-    {
-        var yellow = new SystemConsoleColor { ConsoleColor = ConsoleColor.Yellow };
-        foreach (KeyValuePair<bool, List<(string[], string[])>> groupedModules in modulesWithMissingOptions)
-        {
-            WriteMessage(string.Empty);
-            WriteMessage(groupedModules.Key ? CliCommandStrings.HelpUnavailableOptions : CliCommandStrings.HelpUnavailableExtensionOptions, yellow);
-
-            foreach ((string[] modules, string[] missingOptions) in groupedModules.Value)
-            {
-                if (modules.Length == 0)
-                {
-                    continue;
-                }
-
-                string moduleList = string.Join("\n", modules);
-                StringBuilder line = new();
-                for (int i = 0; i < missingOptions.Length; i++)
-                {
-                    if (i == missingOptions.Length - 1)
-                        line.Append($"--{missingOptions[i]}");
-                    else
-                        line.Append($"--{missingOptions[i]}\n");
-                }
-
-                string format = modules.Length == 1
-                    ? (missingOptions.Length == 1 ? CliCommandStrings.HelpModuleIsMissingTheOptionBelow : CliCommandStrings.HelpModuleIsMissingTheOptionsBelow)
-                    : (missingOptions.Length == 1 ? CliCommandStrings.HelpModulesAreMissingTheOptionBelow : CliCommandStrings.HelpModulesAreMissingTheOptionsBelow);
-                var missing = string.Format(format, moduleList);
-                WriteMessage($"{missing}\n{line}\n");
-            }
-        }
     }
 }
